+++ src/ccbsearch.c
@@ -235,7 +235,7 @@ int scan(result *cres,result *res,FILE *vfp,FILE *ifp, FILE *cifp,DIMS *dims,int
             }
         }else{
             size_t dshape[dims->dims_size];
-            int idx[dims->dims_size];
+            size_t idx[dims->dims_size];
             get_dshape(dshape,dims->shape,dims->dims_size);
             int *offsets=(int *)calloc(cols_size+1,sizeof(int));
 
@@ -277,7 +277,7 @@ int scan(result *cres,result *res,FILE *vfp,FILE *ifp, FILE *cifp,DIMS *dims,int
             }
         }else{
             size_t dshape[dims->dims_size];
-            int idx[dims->dims_size];
+            size_t idx[dims->dims_size];
             get_dshape(dshape,dims->shape,dims->dims_size);
             int *offsets=(int *)calloc(cols_size+1,sizeof(int));
             int j;
